move into happy path test
